agents/coding_agent.py
"""
Autonomous coding agent using CAMEL and DeepSeq
"""
import os
from typing import List, Optional, Dict, Any
from dataclasses import dataclass
import pytest
from langchain.schema import SystemMessage, HumanMessage, AIMessage

@dataclass
class CodingTask:
    description: str
    requirements: List[str]
    test_requirements: Optional[List[str]] = None

class CodingAgent:
    """An autonomous coding agent that can write and test code"""
    
    def __init__(self):
        self.api_key = os.getenv("DEEPSEQ_API_KEY")
        if not self.api_key:
            raise ValueError("DEEPSEQ_API_KEY environment variable not set")
        
        # Will initialize model connection in future PR
        self.model = None
        self.system_message = SystemMessage(content=
            "You are an autonomous coding agent. Your tasks are to:"
            "\n- Write code based on requirements"
            "\n- Create pytest tests"
            "\n- Evaluate code quality"
            "\n- Suggest improvements"
        )
        self.messages = [self.system_message]

    def add_message(self, message: str, role: str = "human") -> None:
        """Add a message to the conversation history"""
        if role == "human":
            self.messages.append(HumanMessage(content=message))
        elif role == "ai":
            self.messages.append(AIMessage(content=message))
        else:
            raise ValueError(f"Invalid role: {role}")

    def create_code(self, task: CodingTask) -> str:
        """Generate code based on task requirements"""
        # Will implement in future PR
        return ""

    def create_tests(self, code: str, task: CodingTask) -> str:
        """Generate pytest tests for the code"""
        # Will implement in future PR
        return ""

    def evaluate_code(self, code: str, tests: str) -> Dict[str, Any]:
        """Evaluate code quality and test results"""
        # Will implement in future PR
        return {}

    def improve_code(self, code: str, evaluation: Dict[str, Any]) -> str:
        """Suggest and implement code improvements"""
        # Will implement in future PR
        return ""
